+++ driver/opentelekomcloud_test.go
@@ -125,8 +125,10 @@ func cleanupResources(driver *Driver) error {
 	if err != nil {
 		return err
 	}
-	if err := driver.client.DeleteFloatingIP(driver.FloatingIP.Value); err != nil {
-		log.Error(err)
+	if driver.FloatingIP.DriverManaged && driver.FloatingIP.Value != "" {
+		if err := driver.client.DeleteFloatingIP(driver.FloatingIP.Value); err != nil {
+			log.Error(err)
+		}
 	}
 	if instanceID != "" {
 		driver.InstanceID = instanceID
@@ -155,6 +157,9 @@ func cleanupResources(driver *Driver) error {
 	}
 	if driver.ManagedSecurityGroupID != "" {
 		_ = driver.client.DeleteSecurityGroup(driver.ManagedSecurityGroupID)
+	}
+	if driver.K8sSecurityGroupID != "" {
+		_ = driver.client.DeleteSecurityGroup(driver.K8sSecurityGroupID)
 	}
 	vpcID, _ := driver.client.FindVPC(vpcName)
 	if vpcID == "" {
